add new test file for cuda tests
