+++ test/onnx/test_pytorch_onnx_onnxruntime.py
@@ -4329,15 +4329,6 @@ class TestONNXRuntime(unittest.TestCase):
         x = torch.randn(2, 4, 5, 6, requires_grad=True)
         self.run_test(GeluModel(), x)
 
-    @unittest.skip("ONNX CI does not build with CUDA")
-    def test_gelu_fp16(self):
-        class GeluModel(torch.nn.Module):
-            def forward(self, x):
-                return torch.nn.functional.gelu(x)
-
-        x = torch.randn(2, 4, 5, 6, requires_grad=True, dtype=torch.float16, device=torch.device('cuda'))
-        self.run_test(GeluModel(), x)
-
     def test_add_inplace(self):
         class InplaceAddModel(torch.nn.Module):
             def forward(self, x):
